Overhaul pt. 1

+++ src/game/entity/entity_transform.h
@@ -1,14 +1,12 @@
 #ifndef ENTITY_TRANSFORM_H
 #define ENTITY_TRANSFORM_H
 
-#include <stdint.h>
-
-#include "game/math/vector3D.h"
+#include "math/vector3D.h"
 
 typedef struct entity_transform_t {
 	vector3D_t position;
 	vector3D_t velocity;
-	uint64_t last_stationary_time;
+	unsigned long long int last_stationary_time;
 } entity_transform_t;
 
 #endif	// ENTITY_TRANSFORM_H
