+++ src/render/vulkan/math/lerp.h
@@ -6,6 +6,7 @@
 // Returns a double-precision floating-point value which is between 0 and 1 when val is between start and end.
 double normalize_delta(const int start, const int end, const int x);
 
-float lerpF(const float start, const float end, const double delta);
+// Linear interpolation with single-precision floating-point numbers.
+float lerpF(const float start, const float end, const float delta);
 
 #endif	// LERP_H
